pyproject.toml
[tool.poetry]
name = "mealplanner"
version = "0.1.0"
description = "Smart Meal Planner - A command-line Python application for meal planning and recipe management"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "mealplanner", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
typer = ">=0.4"
SQLAlchemy = ">=1.4"
alembic = ">=1.8"
pandas = ">=2.0"
python-dotenv = ">=0.21"
tabulate = ">=0.8"
requests = ">=2.0"
rapidfuzz = ">=2.0"
schedule = ">=1.1"
tenacity = ">=8.0"
tqdm = ">=4.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-cov = ">=4.0"

[tool.poetry.scripts]
mealplanner = "mealplanner.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/mealplanner --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]
